Use tags only workflow behavior
Without this, the action assumes there is a release branch that a tag is made on because that is common to most other Grafana repositories.
Those repositories maintain a long-lived release branch to facilitate backports to documentation and code.

Since grafana/oncall doesn't follow this pattern, it doesn't make sense to try and discover the matching release branch.
That action behavior is disabled by the `tags_only: true` setting.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>