Use tags only workflow behavior (#5404)
# What this PR does

Without this, the action assumes there is a release branch that a tag is
made on because that is common to most other Grafana repositories. Those
repositories maintain a long-lived release branch to facilitate
backports to documentation and code.

Since grafana/oncall doesn't follow this pattern, it doesn't make sense
to try and discover the matching release branch. That action behavior is
disabled by the `tags_only: true` setting.

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [x] Documentation added (or `pr:no public docs` PR label added if not
required)
- [x] Added the relevant release notes label (see labels prefixed w/
`release:`). These labels dictate how your PR will
    show up in the autogenerated release notes.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>