Update docker-compose.yml (#5558)
# What this PR does
add mattermost to CELERY_WORKER_QUEUE

# Comment
#5553 
We think that with the 1.16.0 release it would be a good idea to add
mattermost to the CELERY_WORKER_QUEUE so that no one else has the same
issues as we did when we upgraded from 1.9 to 1.16.2 to get the
Mattermost integration via ChatOps.

When we upgraded our OnCall and set up the Mattermost integration, we
couldn't get any messages in MM because Celery was simply ignoring the
Mattermost queue.